+++ R/AdaBoost.R
@@ -108,6 +108,9 @@ fitAdaBoost <- function(population,
   colnames(cvAuc) <- paste0('fold_auc', 1:ncol(cvAuc))
   auc <- unlist(lapply(hyperParamSel, function(x) x$auc))
   
+  cvPrediction <- lapply(hyperParamSel, function(x) x$prediction )
+  cvPrediction <- cvPrediction[[which.max(auc)[1]]]
+  
   hyperSummary <- cbind(do.call(rbind, param), cvAuc, auc= auc)
 
   writeLines('Training Final')
@@ -146,7 +149,8 @@ fitAdaBoost <- function(population,
   
   # return model location (!!!NEED TO ADD CV RESULTS HERE)
   result <- list(model = modelTrained,
-                 trainCVAuc = unlist(cvAuc[bestInd,]),#hyperParamSel,
+                 trainCVAuc = list(value = unlist(cvAuc[bestInd,]),
+                                   prediction = cvPrediction),#hyperParamSel,
                  hyperParamSearch = hyperSummary,
                  modelSettings = list(model = "fitAdaBoost", modelParameters = param.best),
                  metaData = plpData$metaData,
@@ -193,7 +197,7 @@ trainAdaBoost <- function(population, plpData, nEstimators = 50, learningRate = 
 
     auc <- computeAuc(pred)
     writeLines(paste0("CV model obtained CV AUC of ", auc))
-    return(list(auc = auc, aucCV = aucCV))
+    return(list(auc = auc, aucCV = aucCV, prediction = pred[pred$indexes>0,]))
   }
 
   return(result)