+++ R/DecisionTree.R
@@ -132,6 +132,9 @@ fitDecisionTree <- function(population, plpData, param, search='grid', quiet=F,
   colnames(cvAuc) <- paste0('fold_auc', 1:ncol(cvAuc))
   auc <- unlist(lapply(hyperParamSel, function(x) x$auc))
   
+  cvPrediction <- lapply(hyperParamSel, function(x) x$prediction )
+  cvPrediction <- cvPrediction[[which.max(auc)[1]]]
+  
   hyperSummary <- cbind(do.call(rbind, param), cvAuc, auc= auc)
   
   #now train the final model and return coef
@@ -172,7 +175,8 @@ fitDecisionTree <- function(population, plpData, param, search='grid', quiet=F,
   
   # return model location (!!!NEED TO ADD CV RESULTS HERE)
   result <- list(model = modelTrained,
-                 trainCVAuc = unlist(cvAuc[bestInd,]),
+                 trainCVAuc = list(value = unlist(cvAuc[bestInd,]),
+                                   prediction = cvPrediction),
                  hyperParamSearch = hyperSummary,
                  modelSettings = list(model='fitDecisionTree',modelParameters=param.best),
                  metaData = plpData$metaData,
@@ -231,7 +235,7 @@ trainDecisionTree <- function(population, plpData,
     
     aucCV <- lapply(1:max(pred$indexes), function(i){computeAuc(pred[pred$indexes==i,])})
     
-    return(list(auc = auc, aucCV = aucCV))
+    return(list(auc = auc, aucCV = aucCV, prediction = pred[pred$indexes>0,]))
   }
   
   return(result)