+++ R/HelperFunctions.R
@@ -390,7 +390,7 @@ configurePython <- function(envname='PLP', envtype=NULL){
       ParallelLogger::logInfo(paste0('Creating virtual conda environment called ', envname))
       location <- reticulate::conda_create(envname=envname, packages = "python", conda = "auto")
     }
-    packages <- c('numpy','scipy','scikit-learn', 'pandas','pydotplus','keras')
+    packages <- c('numpy','scipy','scikit-learn', 'pandas','pydotplus','keras','joblib')
     ParallelLogger::logInfo(paste0('Adding python dependancies to ', envname))
     reticulate::conda_install(envname=envname, packages = packages, forge = TRUE, pip = FALSE,
                               pip_ignore_installed = TRUE, conda = "auto")