+++ R/RandomForest.R
@@ -149,6 +149,7 @@ fitRandomForest <- function(population, plpData, param, search='grid', quiet=F,
   # run rf_plp for each grid search:
   all_auc <- c()
   all_cvAuc <- c()
+  bestAUC <- 0
   for(i in 1:nrow(param)){
     
     # then run standard python code
@@ -171,12 +172,21 @@ fitRandomForest <- function(population, plpData, param, search='grid', quiet=F,
       
     auc <- computeAuc(pred)
     all_auc <- c(all_auc, auc)
+    
+    
+    if(auc > bestAUC){
+      bestAUC <- auc
+      cvPrediction <- pred[pred$indexes>0,] # need to convert rowId-1?
+    }
+    
     if(!quiet)
       ParallelLogger::logInfo(paste0('Model with settings: ntrees:',param$ntrees[i],' maxDepth: ',param$maxDepth[i], 
                                   'mtry: ', param$mtry[i] , ' obtained AUC of ', auc))
   }
   colnames(all_cvAuc) <- paste0('fold_auc', 1:ncol(all_cvAuc))
   hyperSummary <- cbind(param, all_cvAuc, auc=all_auc)
+  
+  
   
   # now train the final model for the best hyper-parameters previously found
   #reticulate::source_python(system.file(package='PatientLevelPrediction','python','finalRandomForest.py'), envir = e)
@@ -213,7 +223,8 @@ fitRandomForest <- function(population, plpData, param, search='grid', quiet=F,
   
   # return model location
   result <- list(model = modelTrained,
-                 trainCVAuc = all_cvAuc[which.max(all_auc),],
+                 trainCVAuc = list(value = all_cvAuc[which.max(all_auc),],
+                                   prediction = cvPrediction),
                  modelSettings = list(model='randomForest_python',modelParameters=param.best),
                  hyperParamSearch = hyperSummary,
                  metaData = plpData$metaData,