+++ inst/python/naiveBayesFunctions.py
@@ -16,9 +16,9 @@ import math
 from sklearn.naive_bayes import GaussianNB #BernoulliNB
 from scipy.sparse import coo_matrix,csr_matrix,vstack,hstack
 #from sklearn.feature_selection import SelectFromModel#from sklearn.cross_validation import PredefinedSplit
-from sklearn.externals.joblib import Memory
+from joblib import Memory
 #from sklearn.datasets import load_svmlight_file
-from sklearn.externals import joblib
+import joblib
 from sklearn.feature_selection import SelectKBest
 from sklearn.feature_selection import chi2